# proto/gen/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /api/v1/knowledge/search:
        get:
            tags:
                - KnowledgeSearchService
            operationId: KnowledgeSearchService_SearchKnowledge
            parameters:
                - name: query
                  in: query
                  schema:
                    type: string
                - name: topic
                  in: query
                  schema:
                    type: string
                - name: pageSize
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: contentType
                  in: query
                  schema:
                    type: string
                - name: fileFormat
                  in: query
                  schema:
                    type: string
                - name: createdFrom
                  in: query
                  schema:
                    type: string
                - name: createdBefore
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/SearchKnowledgeResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/v1/knowledge/search:rebuild:
        post:
            tags:
                - KnowledgeSearchService
            operationId: KnowledgeSearchService_RebuildKnowledgeSearch
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/RebuildKnowledgeSearchRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/RebuildKnowledgeSearchResponse'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
...
components:
    schemas:
...
        GoogleProtobufAny:
            type: object
            properties:
                '@type':
                    type: string
                    description: The type of the serialized message.
            additionalProperties: true
            description: Contains an arbitrary serialized message along with a @type that describes the type of the serialized message.
...
        KnowledgeSearchResult:
            type: object
            properties:
                document:
                    type: string
                title:
                    type: string
                originalFilename:
                    type: string
                snippet:
                    type: string
                card:
                    type: string
...
        RebuildKnowledgeSearchRequest:
            type: object
            properties: {}
        RebuildKnowledgeSearchResponse:
            type: object
            properties:
                indexedDocuments:
                    type: integer
                    format: int32
                indexedCards:
                    type: integer
                    format: int32
...
        SearchKnowledgeResponse:
            type: object
            properties:
                results:
                    type: array
                    items:
                        $ref: '#/components/schemas/KnowledgeSearchResult'
...
        Status:
            type: object
            properties:
                code:
                    type: integer
                    description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
                    format: int32
                message:
                    type: string
                    description: A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
                details:
                    type: array
                    items:
                        $ref: '#/components/schemas/GoogleProtobufAny'
                    description: A list of messages that carry the error details.  There is a common set of message types for APIs to use.
            description: 'The `Status` type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by [gRPC](https://github.com/grpc). Each `Status` message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the [API Design Guide](https://cloud.google.com/apis/design/errors).'
...